Fixed animation studder in main page (#6)
