+++ src/containers/welcomePage/WelcomePage.tsx
@@ -2,13 +2,8 @@ import React, { useState } from 'react';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles, useTheme, Theme } from '@material-ui/core/styles';
 
-import useScroll from '../../hooks/useScroll';
-import usePaginate from '../../hooks/usePagination';
-
 import SelectionWheel from '../../components/SelectionWheel/SelectionWheel';
-import { firestore } from '../../firebase/init';
-import { getDownloadUrls } from '../../firebase/firestore/firestore';
-import RenderImageGallery from '../../components/RenderImageGallery/RenderImageGallery';
+import RenderWelcomePage from './RenderWelcomePage';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
@@ -33,20 +28,7 @@ const useStyles = makeStyles((theme: Theme) => ({
 const WelcomePage = () => {
   const theme = useTheme();
   const classes = useStyles(theme);
-  const bottomHit = useScroll();
-  const dbRefs = [
-    firestore.collection('images').where('visibility', '==', 'PUBLIC').orderBy('rating', 'desc').orderBy('createDate', 'desc'),
-    firestore.collection('images').where('visibility', '==', 'PUBLIC').orderBy('views', 'desc').orderBy('createDate', 'desc'),
-    firestore.collection('images').where('visibility', '==', 'PUBLIC').orderBy('favorites', 'desc').orderBy('createDate', 'desc'),
-  ];
   const [activeFilter, setActiveFilter] = useState(-1);
-  const [images, , loading, setChangeQuery] = usePaginate(bottomHit, dbRefs[0], ['rating', 'createDate'], 15, getDownloadUrls, true);
-
-  React.useEffect(() => {
-    if (activeFilter === -1) return;
-    setChangeQuery(dbRefs[activeFilter]);
-    // eslint-disable-next-line
-  }, [activeFilter]);
 
   return (
     <div className={classes.root}>
@@ -67,7 +49,7 @@ const WelcomePage = () => {
         </Typography>
       </div>
 
-      <RenderImageGallery images={images} onEmptyMessage={loading ? 'Loading...' : 'Nothing found :/'} />
+      <RenderWelcomePage activeFilter={activeFilter} />
     </div>
   );
 };
